[otbn] Split up insn_yaml.py
This was getting a bit unwieldy: split it into smaller files.
